@@ -3,6 +3,7 @@
   "version": "0.1.0",
   "description": "A simple Node app built on fastify, instantly up and running.",
   "main": "server.js",
+  "type": "module",
   "scripts": {
     "start": "node server.js"
   },
@@ -12,7 +13,7 @@
     "canvas": "^2.11.1"
   },
   "engines": {
-    "node": "14.x"
+    "node": "22.x"
   },
   "repository": {
     "url": "https://glitch.com/edit/#!/glitch-mvp-node"
